Spatial beta diversity for vector

Calculates spatial beta diversity for taxonomic (TD), functional (FD), and phylogenetic (PD) dimensions. Adapted from beta

Usage

spat.beta.vec(x, tree, global = FALSE, spp, nspp, ...)

Arguments

x

A numeric vector with presence-absence data (0 or 1) for a set of species.

tree

It can be a data frame with species traits or a phylogenetic tree.

global

Logical. Mean of pairwise comparisons between focal cell and its neighbors (default) or mean of all pairwise comparisons.

spp

Character. Species names.

nspp

Numeric. Number of species.

...

Additional arguments to be passed passed down from a calling function.

Value

A SpatRaster with beta results (total, replacement, and richness differences).

Details

The TD beta diversity partitioning framework we used was developed by Podani and Schmera (2011) and Carvalho et al. (2012) and expanded to PD and FD by Cardoso et al. (2014).

References

Cardoso, P. et al. 2014. Partitioning taxon, phylogenetic and functional beta diversity into replacement and richness difference components. - Journal of Biogeography 41: 749–761.

Carvalho, J. C. et al. 2012. Determining the relative roles of species replacement and species richness differences in generating beta-diversity patterns. - Global Ecology and Biogeography 21: 760–771.

Podani, J. and Schmera, D. 2011. A new conceptual and methodological framework for exploring and explaining pattern in presence - absence data. - Oikos 120: 1625–1638.