Spatial beta diversity for vector

Spatial beta diversity for vector

Usage

spat.beta.vec(x, tree, global = FALSE, spp, nspp, ...)

Arguments

x

A numeric vector with presence-absence data (0 or 1) for a set of species.

tree

It can be a data frame with species traits or a phylogenetic tree.

global

Logical. Mean of pairwise comparisons between focal cell and its neighbors (default) or mean of all pairwise comparisons.

spp

Character. Species names.

nspp

Numeric. Number of species.

...

Additional arguments to be passed down from a calling function.

Value

A vector with beta results (total, replacement, richness difference, and ratio).